requirements.txt
automate-publishing==1.0
bleach==5.0.1
certifi==2022.6.15
charset-normalizer==2.0.12
commonmark==0.9.1
docutils==0.19
idna==3.3
importlib-metadata==4.12.0
jaraco.classes==3.2.2
keyring==23.9.1
Markdown==3.4.1
more-itertools==8.14.0
pkginfo==1.8.3
Pygments==2.13.0
python-dotenv==0.20.0
readme-renderer==37.1
requests==2.28.0
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
six==1.16.0
twine==4.0.1
urllib3==1.26.9
webencodings==0.5.1
zipp==3.8.1
